# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: SensiLog API
  version: 1.0.0
  description: VALORANT感度・設定管理API
  contact:
    name: SensiLog Team
    url: https://sensilog.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sensilog.com
    description: Production
  - url: http://localhost:3001
    description: Development
...
paths:
...
  /match-data:
    get:
      summary: 試合データ取得
      tags: [MatchData]
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: 開始日
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: 終了日
        - name: maps
          in: query
          schema:
            type: array
            items:
              type: string
          description: マップフィルター
        - name: agents
          in: query
          schema:
            type: array
            items:
              type: string
          description: エージェントフィルター
        - name: gameMode
          in: query
          schema:
            type: string
          description: ゲームモード
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        200:
          description: 試合データ一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/MatchData'
                  total:
                    type: integer
                  hasMore:
                    type: boolean
        401:
          $ref: '#/components/responses/Unauthorized'
  /match-data/sync:
    post:
      summary: 試合データ手動同期
      tags: [MatchData]
      security:
        - bearerAuth: []
      responses:
        202:
          description: 同期開始
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "同期を開始しました"
                  jobId:
                    type: string
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          description: レート制限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    MatchData:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        matchId:
          type: string
          description: Riot Match ID
        gameStartTime:
          type: string
          format: date-time
        gameEndTime:
          type: string
          format: date-time
        mapName:
          type: string
          description: マップ名
        gameMode:
          type: string
          description: ゲームモード
        agentName:
          type: string
          description: 使用エージェント
        kills:
          type: integer
          minimum: 0
        deaths:
          type: integer
          minimum: 0
        assists:
          type: integer
          minimum: 0
        combatScore:
          type: number
          format: float
          minimum: 0
          description: ACS（Average Combat Score）
        headshotPercentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: ヘッドショット率
        kdRatio:
          type: number
          format: float
          minimum: 0
          description: K/D比
        adr:
          type: number
          format: float
          minimum: 0
          description: 平均ダメージ/ラウンド
        damageDealt:
          type: integer
          minimum: 0
        roundsPlayed:
          type: integer
          minimum: 0
        teamWon:
          type: boolean
        rankTier:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - matchId
        - gameStartTime
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
        code:
          type: string
          description: エラーコード
        details:
          type: object
          description: 詳細情報
      required:
        - error
  responses:
...
    Unauthorized:
      description: 認証が必要
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT